Commit after running angular-frontend from the express-backend
Updates:
========
Code for running the frontend from the express backend added to the
backend app. This was done because on firefox the CORS and allow-origin
type of issues were surfacing which were quite annoying.

Other changes:
==============
Some untracked files from the previous revision were not committed
because there was a file "selfsigned.key" which was locked by ubuntu and
couldnot be staged to commit and it prevented other files as well
because i was selecting all files and staging. This file is still not
committed because due to some problems with it (as locked) i cannot
stage it.